+++ lgbt/lgbt/lgbt.py
@@ -3,210 +3,8 @@ import sys
 import inspect
 import os
 
-from bar import DynemicBar, AdvancedBar
-from basicobjects import LegacyBar
-
-class Tracker:
-	def __init__(self, number=None):
-		if type(number) in [float, int, type(None)]:
-			self._number = number
-		elif type(number) == type(self):
-			self = number
-		else:
-			raise ValueError("Not correct type of variable")
-
-	@property
-	def item(self):
-		return self._number
-	
-	@item.setter
-	def item(self, value):
-		if type(value) in [float, int, type(None)]:
-			self._number = value
-		elif type(value) == type(self):
-			self = value
-		else:
-			raise ValueError("Not correct type of variable")
-
-	
-	def __str__(self):
-		return str(self._number)
-	
-	def __add__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number + value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number + value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __sub__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number - value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number - value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __mul__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number * value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number * value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __truediv__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number / value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number / value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __floordiv__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number // value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number // value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __mod__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number % value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number % value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __pow__(self, value):
-		if type(value) == type(self):
-			return Tracker(self._number ** value.item())
-		elif type(value) in [float, int]:
-			return Tracker(self._number ** value)
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __eq__(self, value):
-		if type(value) == type(self):
-			return self._number == value.item()
-		elif type(value) in [float, int, type(None)]:
-			return self._number == value
-		else:
-			raise ValueError("Not correct type of variable")
-
-	def __ne__(self, value):
-		if type(value) == type(self):
-			return self._number != value.item()
-		elif type(value) in [float, int, type(None)]:
-			return self._number != value
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __lt__(self, value):
-		if type(value) == type(self):
-			return self._number < value.item()
-		elif type(value) in [float, int]:
-			return self._number < value
-		else:
-			raise ValueError("Not correct type of variable")
-
-	def __gt__(self, value):
-		if type(value) == type(self):
-			return self._number > value.item()
-		elif type(value) in [float, int]:
-			return self._number > value
-		else:
-			raise ValueError("Not correct type of variable")
-	
-	def __le__(self, value):
-		if type(value) == type(self):
-			return self._number <= value.item()
-		elif type(value) in [float, int]:
-			return self._number <= value
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __ge__(self, value):
-		if type(value) == type(self):
-			return self._number >= value.item()
-		elif type(value) in [float, int]:
-			return self._number >= value
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __iadd__(self, value):
-		if type(value) == type(self):
-			self._number += value.item()
-			return self
-		elif type(value) in [float, int]:
-			self._number += value
-			return self
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __isub__(self, value):
-		if type(value) == type(self):
-			self._number -= value.item()
-			return self
-		elif type(value) in [float, int]:
-			self._number -= value
-			return self
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __imul__(self, value):
-		if type(value) == type(self):
-			self._number *= value.item()
-			return self
-		elif type(value) in [float, int]:
-			self._number *= value
-			return self
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __itruediv__(self, value):
-		if type(value) == type(self):
-			self._number /= value.item()
-			return self
-		elif type(value) in [float, int]:
-			self._number /= value
-			return self
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __ifloordiv__(self, value):
-		if type(value) == type(self):
-			self._number //= value.item()
-			return self
-		elif type(value) in [float, int]:
-			self._number //= value
-			return self
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __ipow__(self, value):
-		if type(value) == type(self):
-			self._number **= value.item()
-			return self
-		elif type(value) in [float, int]:
-			self._number **= value
-			return self
-		else:
-			raise ValueError("Not correct type of variable")
-		
-	def __pos__(self):
-		self._number = +self._number
-		return self
-	
-	def __neg__(self):
-		self._number = -self._number
-		return self
-
-	def __abs__(self):
-		return +self
-
+from .bar import DynemicBar, AdvancedBar
+from .basicobjects import Tracker
 
 
 class lgbt():
@@ -232,7 +30,7 @@ class lgbt():
 		self._minintervals = minintervals
 		self._current_iter = 0
 		self._is_end = False
-
+		self._bar = DynemicBar(total=self._total, hero=hero, desc=desc, mode=mode)
 		self._miniters = max(1, round(self._total/self._miniters))
 
 	def __init__tracker__(self, iterable=None, total=None, desc="", miniters=2500, minintervals=0.1, hero='rainbow', mode='default', tracker=None):
